+++ src/dune_rules/coq/coq_stanza.mli
@@ -31,6 +31,7 @@ module Theory : sig
     ; project : Dune_project.t
     ; synopsis : string option
     ; modules : Ordered_set_lang.t
+    ; load_only_modules : Ordered_set_lang.t
     ; boot : bool
     ; enabled_if : Blang.t
     ; buildable : Buildable.t