+++ test/blackbox-tests/test-cases/coq/load-only.t/D.v
@@ -1,0 +1,3 @@
+Load A.
+Load "B/C.v".
+Load E.